requirements.txt
pandas==2.2.2
flask_restful==0.3.10
Flask==3.0.0
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy<1.24
reportlab==4.2.0
SQLAlchemy==2.0.25
pyyaml
openpyxl